Use uint64 for indirect pin refcounts
Platform-dependent behavior is not nice, and negative refcounts are
not very useful.

License: MIT

+++ core/commands/pin.go
@@ -275,7 +275,7 @@ Example:
 
 type RefKeyObject struct {
 	Type  string
-	Count int
+	Count uint64
 }
 
 type RefKeyList struct {
